Moved old test into integration tests
